* 'implement_setting_screen' of github.com-tony:nustechnology/Transcribe-Summarize-ClearHear:
  Fix code.
  Format code.
  Resolved some feedback.
  In-Session Font Size Adjustments ( A+ / A- )
  Fix some bugs.
  Updated README.md file.
  Fixed library issue.
  Resolved some feedback.
  Live session controls (Start, Pause, Resume, and Stop)

# Conflicts:
#	lib/screen/settings/bindings/settings_binding.dart
#	lib/screen/settings/controllers/settings_controller.dart
#	lib/screen/settings/widgets/settings_section_card.dart
#	lib/shared/models/settings_model.dart

+++ lib/shared/models/settings_model.dart
@@ -57,7 +57,8 @@ class SettingsModel {
     return SettingsModel(
       fontSize: ((map['font_size'] as num?)?.toDouble() ??
               CaptionSizeConfig.defaultSize)
-          .clamp(CaptionSizeConfig.min, CaptionSizeConfig.max),
+          .clamp(CaptionSizeConfig.min, CaptionSizeConfig.max)
+          .toDouble(),
       theme: map['theme'] as String? ?? 'system',
       savingEnabled: (map['saving_enabled'] as int? ?? 1) == 1,
       keepScreenOn: (map['keep_screen_on'] as int? ?? 0) == 1,
